lint: fix linter issues on Darwin platform
