+++ ios/parselsblk_internal_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // Package ios is a collection of interfaces to the local storage subsystem;
 // the package includes OS-dependent implementations for those interfaces.
 /*
